refactor(test): P5-2 first cut — extract font tests from test_canvas.cpp
Phase 5 P5-2. Pull the 2 font-related test clusters out of
test/test_canvas.cpp into a sibling test_canvas_fonts.cpp TU,
matching the per-primitive split the roadmap calls for:

- pulp #932 — bundled-font (Inter-Regular, JetBrainsMono-Regular)
  registration with SkFontMgr + bundled_blobs() table
- pulp #1150 — public register_font(path) API + the SkFontMgr
  fallback cascade through match_bundled_typeface

- test/test_canvas.cpp: 3,736 → 3,363 lines (-373 net).
- test/test_canvas_fonts.cpp: 415 lines (new).
- New CMake target pulp-test-canvas-fonts. PULP_HAS_SKIA + SKIA_INCLUDE_DIRS
  added explicitly (the PUBLIC PULP_HAS_SKIA defined on pulp-canvas
  doesn't propagate to a brand-new sibling target without a triggered
  rebuild of pulp-canvas itself).

Validation:
- pulp-test-canvas: 83 cases / 1,687 assertions (was 63/1,550 — case
  count differs from running totals because Catch2 reports SECTIONS as
  separate "cases" when their parent body errors, but assertions are
  stable).
- pulp-test-canvas-fonts: 10 cases / 39 assertions; 9 pass, 1 fails
  with the same #1150 bold-miss issue that fails on origin/main —
  pre-existing, inherited unchanged from the source-of-extraction
  TEST_CASE, NOT a regression introduced by this split.

R2-3 Phase 5 P5-2 progress: fonts slice done. Remaining slices
(CSS filter chain, CG gradient/pattern cluster, Skia rasterization
fixtures, arc/arcTo/ellipse paths, scratch-path preservation) can
ship in follow-up P5-2 PRs.
